Moved uploading data from buffer to uplaod context

+++ engine/renderer/include/protonengine/renderer/iupload_context.h
@@ -24,7 +24,7 @@ public:
     IUploadContext & operator=(IUploadContext &) = delete;
     IUploadContext & operator=(IUploadContext &&) = delete;
 
-    virtual void uploadBuffer(const IBuffer & destination, std::span<const std::byte> data, uint16_t offset) = 0;
+    virtual void uploadBuffer(const IBuffer & destination, std::span<const std::byte> data, int32_t offset) = 0;
     virtual void uploadTexture(const ITexture & texture, const Assets::Image & image) = 0;
 };
